requirements-dev.txt
# =============================================================================
# requirements-dev.txt — Development and quality tools for MyNotion
# =============================================================================
# Install: pip install -r requirements-dev.txt
# These are NOT needed to run the app, only for development.
# =============================================================================

# ── Linting & formatting ─────────────────────────────────────────────────────
# Ruff replaces black, isort, flake8, pyflakes — written in Rust, extremely fast
ruff>=0.9.0

# ── Type checking ─────────────────────────────────────────────────────────────
mypy>=1.14.0
PyQt6-stubs                  # Type stubs so mypy understands PyQt6 classes

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.0
pytest-qt>=4.4.0             # qtbot fixture for PyQt6 widget testing
pytest-asyncio>=0.24.0       # async test support
pytest-cov>=6.0              # coverage reporting (optional but useful)

# ── Async Qt integration ─────────────────────────────────────────────────────
qasync>=0.27.0               # Bridges asyncio event loop with Qt event loop

# ── Git hooks ─────────────────────────────────────────────────────────────────
pre-commit>=4.0.0
